requirements.txt
aniso8601==9.0.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
coverage==6.2
cryptography==36.0.0
Flask==2.0.2
Flask-JWT-Extended==4.3.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.28
SQLAlchemy-serializer==1.4.1
Werkzeug==2.0.2

flasgger~=0.9.5
argparse~=1.4.0
python-dateutil~=2.8.2